refactor: moved insertData to withLinks plugin
