+++ packages/core/admin/admin/src/content-manager/components/BlocksEditor/plugins/index.js
@@ -1,5 +1,4 @@
-import { withDataInsertion } from './withDataInsertion';
 import { withLinks } from './withLinks';
 import { withStrapiSchema } from './withStrapiSchema';
 
-export { withLinks, withStrapiSchema, withDataInsertion };
+export { withLinks, withStrapiSchema };
